Fix crash sending "count" method to NSNull, inside [QLKCue hasChildren]
I think what triggered this is:
If a workspace has a passcode and the user gets it wrong, [QLKWorkspace disconnect] is called. This method sets the root cue's QLKOSCCuesKey to NSNull.
If you ask hasChildren of its root cue before it finished connecting successfully (ie. as part of a NSTableView data source or something), you crash.